# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для блога с аутентификацией",
        "title": "Blog API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Роли пользователей и аутентификация"
                ],
                "summary": "Залогинить пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LoginUserResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh-token": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Роли пользователей и аутентификация"
                ],
                "summary": "Обновить токен пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RefreshUserTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RefreshUserTokenResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Роли пользователей и аутентификация"
                ],
                "summary": "Зарегистрировать пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegistrateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RegistrateUserResponse"
                        }
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Просмотр постов"
                ],
                "summary": "Просмотр постов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен авторизации",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetPostsResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Управление постами"
                ],
                "summary": "Создать пост",
                "parameters": [
                    {
                        "description": "Данные поста",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Токен авторизации",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePostResponse"
                        }
                    }
                }
            }
        },
        "/api/posts/{postId}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Управление постами"
                ],
                "summary": "Редактировать пост",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID поста",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое название поста",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EditPostRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Токен авторизации",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EditPostResponse"
                        }
                    }
                }
            }
        },
        "/api/posts/{postId}/images": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Управление постами"
                ],
                "summary": "Добавить картинку к посту",
                "parameters": [
                    {
                        "type": "string",
                        "name": "postId",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Токен авторизации",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AddImageToPostResponse"
                        }
                    }
                }
            }
        },
        "/api/posts/{postId}/images/{imageId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Управление постами"
                ],
                "summary": "Удалить картинку из поста",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID поста",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID картинки",
                        "name": "imageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Токен авторизации",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteImageFromPostResponse"
                        }
                    }
                }
            }
        },
        "/api/posts/{postId}/status": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Управление постами"
                ],
                "summary": "Опубликовать пост",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID поста",
                        "name": "postId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Статус поста",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PublishPostRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Токен авторизации",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PublishPostResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddImageToPostResponse": {
            "type": "object",
            "properties": {
                "post_id": {
                    "type": "string"
                }
            }
        },
        "models.CreatePostRequest": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.CreatePostResponse": {
            "type": "object",
            "properties": {
                "post_id": {
                    "type": "string"
                }
            }
        },
        "models.DeleteImageFromPostResponse": {
            "type": "object",
            "properties": {
                "post_id": {
                    "type": "string"
                }
            }
        },
        "models.EditPostRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.EditPostResponse": {
            "type": "object",
            "properties": {
                "post_id": {
                    "type": "string"
                }
            }
        },
        "models.GetPostsResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Post"
                    }
                }
            }
        },
        "models.LoginUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.LoginUserResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Post": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "post_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.PublishPostRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "models.PublishPostResponse": {
            "type": "object",
            "properties": {
                "post_id": {
                    "type": "string"
                }
            }
        },
        "models.RefreshUserTokenRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "models.RefreshUserTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "models.RegistrateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "models.RegistrateUserResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        }
    }
}